requirements.txt
streamlit
yfinance
pandas
numpy
plotly
scikit-learn
requests
fastapi
uvicorn
python-dotenv
